requirements.txt
Django==1.7.1
Pillow==2.6.1
django-annoying==0.8.0
django-nose==1.2
google-api-python-client==1.3.1
httplib2==0.9
mock==1.0.1
nose==1.3.4
oauth2client==1.3.2
psycopg2==2.5.4
pyasn1==0.1.7
pyasn1-modules==0.0.5
rsa==3.1.4
simplejson==3.6.5
uritemplate==0.6
wsgiref==0.1.2
requests==2.4.3
